pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "sscbt-experiments"
version = "0.1.0"
description = "Synthetic experiments for shift-set contextual Bradley-Terry ranking."
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scipy",
  "pandas",
  "matplotlib",
  "pyyaml",
]
